+++ props/list_style_position.go
@@ -1,0 +1,12 @@
+package props
+
+type ListStylePosition string
+
+const (
+	ListStylePositionInside  ListStylePosition = "inside"
+	ListStylePositionOutside ListStylePosition = "outside"
+)
+
+func (l ListStylePosition) String() string {
+	return string(l)
+}